requirements.txt
# Core dependencies
torch>=2.0.0
stable-baselines3>=2.0.0
gymnasium>=0.29.0
numpy>=1.21.0
pyyaml>=6.0

# Environment
crafter>=1.8.0

# Logging and monitoring
tensorboard>=2.10.0
wandb>=0.15.0

# LLM integration
requests>=2.28.0

# Development (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0 